requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
elasticsearch==8.11.0
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
numpy==1.25.2
openai==1.3.7